Change the netboot configuration back to lan0 with udev rules.
The udev rules change is tuned for one particular computer where
the MAC addresses of the 2 ethernet ports are ending in 5 and 9
respectively.

IHMO this rules files should be somehow generated for each robot
with an exact mapping, rather than trying to be generic.

The current approach is leading to random renames and the netboot
configuration will only work if eth1 is not renamed.

Another option would be to remove the udev rules file completely and
let c2 use eth0/eth1 everywhere (keeping biosname=0 in the kernel
options).

Signed-off-by: Matthieu Herrb <matthieu.herrb@laas.fr>